% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastLmX.R
\name{fastLmX}
\alias{fastLmX}
\title{Rewritten weighted Linear regression using
RcppArmadillo}
\usage{
fastLmX(X, ...)
}
\arguments{
\item{X}{a matrix for all explantory variables}

\item{...}{other argument such as a colVector for output variable and a colvector for weights}
}
\value{
fastLmX object that includes coef of the reggression results
}
\description{
Rewritten weighted Linear regression using
RcppArmadillo
}
