% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart_lr.R
\name{smart_Pure}
\alias{smart_Pure}
\title{smart Pure method of Linear Regression that automatically select
 Armadillo method or lm.wfit method based on
 dimension of the dataset}
\usage{
smart_Pure(X, Y, Z)
}
\arguments{
\item{X}{a matrix for all explantory variables}

\item{Y}{a vector for output variables}

\item{Z}{a vector for weight}
}
\description{
smart Pure method of Linear Regression that automatically select
 Armadillo method or lm.wfit method based on
 dimension of the dataset
}
